requirements.txt
# Core dependencies for Context OS

# Configuration management
pyyaml>=6.0.1

# Clipboard monitoring (for Phase 2) - Native macOS integration
pyobjc-core>=11.0
pyobjc-framework-Cocoa>=11.0
pyobjc-framework-UserNotifications>=11.0,<12.0  # For native macOS notifications

# LLM integration (for Phase 3)
openai>=1.12.0

# GUI framework (for Phase 4) - PyQt5 for macOS 12 compatibility
PyQt5>=5.15.0

# Additional utilities
python-dateutil>=2.8.2

# Markdown rendering (for UI)
mistune>=3.0.0

# Symbolic mathematics (for calculator tool)
sympy>=1.12

# Image processing (for clipboard image support)
Pillow>=10.0.0

# Screenshot capture (for screenshot stream adapter)
mss>=9.0.0

# Image hashing for deduplication
imagehash>=4.3.0
